+++ pkg/local_object_storage/engine/container_test.go
@@ -28,7 +28,9 @@ func TestStorageEngine_ContainerCleanUp(t *testing.T) {
 	path := t.TempDir()
 
 	synctest.Test(t, func(t *testing.T) {
-		e := New(WithContainersSource(cnrSource{}))
+		e := New(
+			WithContainersSource(cnrSource{}),
+		)
 		t.Cleanup(func() {
 			_ = e.Close()
 		})