+++ pkg/local_object_storage/engine/engine.go
@@ -5,7 +5,6 @@ import (
 	"io"
 	"sync"
 	"sync/atomic"
-	"time"
 
 	iec "github.com/nspcc-dev/neofs-node/internal/ec"
 	containercore "github.com/nspcc-dev/neofs-node/pkg/core/container"
@@ -57,18 +56,10 @@ type shardInterface interface {
 	ReadECPartHeader(cid.ID, oid.ID, iec.PartInfo, []byte) (int, error)
 }
 
-type putTask struct {
-	addr   oid.Address
-	obj    *object.Object
-	objBin []byte
-	retCh  chan error
-}
-
 type shardWrapper struct {
 	errorCount *atomic.Uint32
 	*shard.Shard
 	shardIface shardInterface // TODO: make Shard a shardInterface
-	putCh      chan putTask
 	engine     *StorageEngine
 }
 
@@ -228,9 +219,6 @@ type cfg struct {
 
 	metrics MetricRegister
 
-	objectPutTimeout time.Duration
-	shardPoolSize    uint32
-
 	containerSource containercore.Source
 
 	isIgnoreUninitedShards bool
@@ -239,8 +227,6 @@ type cfg struct {
 func defaultCfg() *cfg {
 	return &cfg{
 		log: zap.L(),
-
-		shardPoolSize: 20,
 	}
 }
 
@@ -276,13 +262,6 @@ func WithMetrics(v MetricRegister) Option {
 	}
 }
 
-// WithShardPoolSize returns option to specify size of worker pool for each shard.
-func WithShardPoolSize(sz uint32) Option {
-	return func(c *cfg) {
-		c.shardPoolSize = sz
-	}
-}
-
 // WithErrorThreshold returns an option to specify size amount of errors after which
 // shard is moved to read-only mode.
 func WithErrorThreshold(sz uint32) Option {
@@ -302,14 +281,5 @@ func WithContainersSource(cs containercore.Source) Option {
 func WithIgnoreUninitedShards(flag bool) Option {
 	return func(c *cfg) {
 		c.isIgnoreUninitedShards = flag
-	}
-}
-
-// WithObjectPutRetryTimeout return an option to specify time for object PUT operation.
-// It does not stop any disk operation, only affects retryes policy. Zero value
-// is acceptable and means no retry on any shard.
-func WithObjectPutRetryTimeout(t time.Duration) Option {
-	return func(c *cfg) {
-		c.objectPutTimeout = t
 	}
 }