+++ pkg/local_object_storage/engine/evacuate.go
@@ -115,16 +115,16 @@ mainLoop:
 					if _, ok := shardMap[shards[j].ID().String()]; ok {
 						continue
 					}
-					err = e.putToShard(ctx, shards[j], addr, obj, nil)
-					if err == nil || errors.Is(err, errExists) {
-						if !errors.Is(err, errExists) {
-							e.log.Debug("object is moved to another shard",
-								zap.String("from", sidList[n]),
-								zap.Stringer("to", shards[j].ID()),
-								zap.Stringer("addr", addr))
-
-							count++
-						}
+					err = e.putToShard(shards[j], addr, obj, nil)
+					if err == nil {
+						e.log.Debug("object is moved to another shard",
+							zap.String("from", sidList[n]),
+							zap.Stringer("to", shards[j].ID()),
+							zap.Stringer("addr", addr))
+						count++
+						continue loop
+					}
+					if errors.Is(err, errExists) {
 						continue loop
 					}