+++ CHANGELOG.md
@@ -20,10 +20,15 @@ Changelog for NeoFS Node
 - Proxied GET data is not checked against SHA256 hash from header (#4081)
 - SN requests EC part ranges using GET API instead of deprecated RANGE one (#4067)
 - In-container SN no longer attaches session/bearer tokens to spawned GET/HEAD requests (#4067)
+- SN no longer limits number of concurrent outgoing requests in PUT handler (#4092)
+- SN no longer limits number of concurrent shard PUT ops (#4092)
+- SN no longer limits local PUT with static timeout (#4092)
+- SN no longer limits number of concurrent outgoing requests in SearchV2 handler (#4092)
 
 ### Removed
 - Compression support from FSTree (#4054)
 - IR container structurization migration (#4062)
+- ADM no longer adds `storage.shard_pool_size` to produced SN configurations (#XXXK)
 
 ### Updated
 - `github.com/nspcc-dev/neofs-sdk-go` module to `v1.0.0-rc.20.0.20260703200507-4f450009c764` (#4065)
@@ -45,6 +50,12 @@ IR container structurization migration introduced in v0.51.0 has been removed.
 Inner Ring nodes updating from versions earlier than v0.51.0 must first update
 to a v0.51.0-v0.54.0 release and let the migration complete before updating to
 v0.55.0.
+
+Drop no longer supported options from all SN configurations:
+- `object.put.pool_size_remote`
+- `storage.shard_pool_size`
+- `storage.put_retry_timeout`
+- `object.search.pool_size`
 
 ## [0.54.0] - 2026-06-29 - Ando